Création du RTOController
